# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for a web app that allows people to manage and easily find/search\ncontents for their boxes without having to manually open boxes.",
        "title": "Boom Box API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/api/v1",
    "paths": {
        "/{user_id}/add-box": {
            "post": {
                "tags": [
                    "Boxes"
                ],
                "summary": "Add a box to a user's collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Unique user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request: (error description)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{user_id}/boxes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Boxes"
                ],
                "summary": "Get all boxes from a user based off of the user id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Unique user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request: (error description)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Box": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "boxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Box"
                    }
                }
            }
        }
    }
}